Some architectural changes
Created an interface, TabletHostingServer, that both TabletServer and ScanServer implement.
ScanServer no longer extends TabletServer, which means that we can back out some of the changes
that we made to TabletServer early in feature development. Backed out the constructor changes
to Tablet and created ScanServer.ReadOnlyTablet which overrides methods called in the constructor
that set up the WAL and modify metadata to do nothing.