Move Hadoop Trash handling to Volume
Move Trash handling from GCRun and SimpleGarbageCollector to
Volume. The benefit of this code reorganization is a single
code path for dealing with Hadoop Trash and creating a single
Trash object per volume instead of creating a new Trash object
every time we want to move a file to the Trash.